+++ src/view/app/Welcome/index.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import { Welcome } from './welcome';
+
+ReactDOM.render(<Welcome />, document.getElementById('root'));